refactor(connections): extract resource and health journeys
